+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Unreleased
 - Skips parsing when response 304 Not Modified.  
   This fixes fetching from servers that send a Content-Type along with 304.
+- Skips parsing when response 204 No Content.  
+  A fail-safe similar to the above.
 
 ## 0.1.337 (Dec 11, 2015)
 - Scoped release for personal use and profit.